+++ src/main/java/de/marhali/easyi18n/io/parser/json5/Json5ArrayMapper.java
@@ -6,7 +6,7 @@ import de.marhali.json5.Json5;
 import de.marhali.json5.Json5Array;
 import de.marhali.json5.Json5Primitive;
 
-import org.apache.commons.lang.math.NumberUtils;
+import org.apache.commons.lang3.math.NumberUtils;
 
 import java.io.IOException;
 
@@ -40,7 +40,7 @@ public class Json5ArrayMapper extends ArrayMapper {
             } else {
                 if(StringUtil.isHexString(element)) {
                     array.add(Json5Primitive.of(element, true));
-                } else if(NumberUtils.isNumber(element)) {
+                } else if(NumberUtils.isCreatable(element)) {
                     array.add(Json5Primitive.of(NumberUtils.createNumber(element)));
                 } else {
                     array.add(Json5Primitive.of(element));